modified long long to long

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: lpaiva <lpaiva@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/07 21:27:51 by lpaiva            #+#    #+#             */
-/*   Updated: 2025/11/10 18:32:32 by lpaiva           ###   ########.fr       */
+/*   Updated: 2025/11/12 22:33:26 by lpaiva           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,8 +23,8 @@
 int	ft_putchar(char c, int count);
 int	ft_putstr(char *str, int count);
 int	ft_printf(const char *format, ...);
-int	ft_putnbr_base(long long nbr, char *base, int base_len, int count);
-int	ft_putnbr_base_uns(unsigned long long nbr, char *base, int base_len,
+int	ft_putnbr_base(long nbr, char *base, int base_len, int count);
+int	ft_putnbr_base_uns(unsigned long nbr, char *base, int base_len,
 		int count);
 int	ft_ptptr(void *ptr, int count);
